refactor: extract scroll container into new component and add semantic html tags
